同步Develop

+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/crypto/PKCS7Encoder.java
@@ -47,7 +47,7 @@ public class PKCS7Encoder {
    * @return 删除补位字符后的明文
    */
   public static byte[] decode(byte[] decrypted) {
-    int pad = (int) decrypted[decrypted.length - 1];
+    int pad = decrypted[decrypted.length - 1];
     if (pad < 1 || pad > 32) {
       pad = 0;
     }
